refactor: Score 패키지 재이동 및 Finished 상태 추가
