+++ src/main/java/bowling/domain/frame/LastFrame.java
@@ -8,7 +8,6 @@ import java.util.List;
 import bowling.domain.BowlingGame;
 import bowling.domain.BowlingGameFrameRecord;
 import bowling.domain.frame.state.Ready;
-import bowling.domain.Score;
 import bowling.domain.frame.state.State;
 
 public class LastFrame implements Frame {
