+++ src/main/java/bowling/domain/frame/state/Finished.java
@@ -1,0 +1,13 @@
+package bowling.domain.frame.state;
+
+abstract class Finished implements State {
+    @Override
+    public final State bowl(int pins) {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public final boolean isFinish() {
+        return true;
+    }
+}
