+++ src/main/java/bowling/domain/frame/state/Miss.java
@@ -2,9 +2,9 @@ package bowling.domain.frame.state;
 
 import java.util.List;
 
-import bowling.domain.Score;
+import bowling.domain.frame.Score;
 
-public class Miss implements State {
+public class Miss extends Finished {
     private static final String INVALID_PINS_EXCEPTION_MESSAGE = "Miss의 조건을 만족하지 않습니다.";
 
     private final Pins firstPins;
@@ -17,18 +17,8 @@ public class Miss implements State {
     }
 
     @Override
-    public State bowl(int pins) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
     public Score createScore() {
         return new Score(List.of(firstPins.getPins(), secondPins.getPins()));
-    }
-
-    @Override
-    public boolean isFinish() {
-        return true;
     }
 
     @Override
