+++ src/main/java/bowling/domain/frame/state/Ready.java
@@ -2,7 +2,7 @@ package bowling.domain.frame.state;
 
 import java.util.List;
 
-import bowling.domain.Score;
+import bowling.domain.frame.Score;
 
 public class Ready implements State {
     @Override
