+++ src/main/java/bowling/domain/frame/state/Spare.java
@@ -2,9 +2,9 @@ package bowling.domain.frame.state;
 
 import java.util.List;
 
-import bowling.domain.Score;
+import bowling.domain.frame.Score;
 
-public class Spare implements State {
+public class Spare extends Finished {
     private static final String INVALID_PINS_EXCEPTION_MESSAGE = "스페어의 조건을 만족하지 않습니다.";
 
     private final Pins firstPins;
@@ -17,18 +17,8 @@ public class Spare implements State {
     }
 
     @Override
-    public State bowl(int pins) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
     public Score createScore() {
         return new Score(List.of(firstPins.getPins(), secondPins.getPins()), true);
-    }
-
-    @Override
-    public boolean isFinish() {
-        return true;
     }
 
     @Override
