+++ src/main/java/bowling/domain/frame/state/State.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame.state;
 
-import bowling.domain.Score;
+import bowling.domain.frame.Score;
 
 public interface State {
     State bowl(int pins);
