+++ src/main/java/bowling/domain/frame/state/Strike.java
@@ -2,22 +2,12 @@ package bowling.domain.frame.state;
 
 import java.util.List;
 
-import bowling.domain.Score;
+import bowling.domain.frame.Score;
 
-public class Strike implements State {
-    @Override
-    public State bowl(int pins) {
-        throw new UnsupportedOperationException();
-    }
-
+public class Strike extends Finished {
     @Override
     public Score createScore() {
         return new Score(List.of(10));
-    }
-
-    @Override
-    public boolean isFinish() {
-        return true;
     }
 
     @Override
